Remove last_update to save on database space
sensor.playlist_sensor fills up the HA database due to the last_update attribute changing every minute. The list of playlists does not change very frequently so this feels unnecessary to pepper the database with the info. 

Alternatives, the update frequency could be lowered for the playlist sensor. Say hourly updates. Or the last_update attribute could be made optional via the config flow options. I'm happy to write the PR for these if one solution is preferred.